@@ -6,7 +6,7 @@
 
     <groupId>me.cbhud</groupId>
     <artifactId>CastleSiege</artifactId>
-    <version>2.0.4</version>
+    <version>2.0.6</version>
     <packaging>jar</packaging>
 
     <name>CastleSiege</name>
@@ -24,8 +24,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.1</version>
                 <configuration>
-                    <source>15</source>
-                    <target>15</target>
+                    <source>16</source>
+                    <target>16</target>
                 </configuration>
             </plugin>
 
