@@ -4,9 +4,7 @@ import me.cbhud.castlesiege.CastleSiege;
 import me.cbhud.castlesiege.arena.Arena;
 import me.cbhud.castlesiege.player.PlayerState;
 import me.cbhud.castlesiege.team.Team;
-import org.bukkit.Location;
-import org.bukkit.Material;
-import org.bukkit.World;
+import org.bukkit.*;
 import org.bukkit.attribute.Attribute;
 import org.bukkit.entity.*;
 import org.bukkit.event.EventHandler;
@@ -15,6 +13,9 @@ import org.bukkit.event.entity.EntityDamageByEntityEvent;
 
 import org.bukkit.event.entity.EntityDeathEvent;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.FireworkMeta;
+
+import java.util.concurrent.ThreadLocalRandom;
 
 public class MobManager implements Listener {
 
@@ -50,7 +51,7 @@ public class MobManager implements Listener {
         kingZombie.getAttribute(Attribute.GENERIC_MAX_HEALTH).setBaseValue(maxHealth);
         kingZombie.setHealth(maxHealth);
 
-        kingZombie.getEquipment().setHelmet(new ItemStack(Material.CARVED_PUMPKIN));
+        kingZombie.getEquipment().setHelmet(new ItemStack(Material.GOLDEN_HELMET));
     }
 
     public double getZombieHealth(Zombie zombie) {
@@ -71,14 +72,6 @@ public class MobManager implements Listener {
             }
         }
         return null;
-    }
-
-    public void removeCustomZombie(Arena arena) {
-        for (LivingEntity entity : arena.getKingSpawn().getWorld().getLivingEntities()) {
-            if (entity instanceof Zombie && entity.getCustomName() != null && entity.getCustomName().contains("King")) {
-                entity.remove();
-            }
-        }
     }
 
 
@@ -128,13 +121,131 @@ public class MobManager implements Listener {
         }
     }
 
+    public void removeCustomZombie(Arena arena) {
+        Location fireworkLoc = arena.getKingSpawn(); // fallback
+
+        for (LivingEntity entity : arena.getKingSpawn().getWorld().getLivingEntities()) {
+            if (entity instanceof Zombie && entity.getCustomName() != null && entity.getCustomName().contains("King")) {
+                fireworkLoc = entity.getLocation().clone();
+                entity.remove();
+                break; // only one king
+            }
+        }
+
+        // 🎆 Celebrate defenders win (3–4 seconds)
+        spawnWinFireworks(fireworkLoc);
+    }
+
+    private void spawnWinFireworks(Location loc) {
+        if (loc == null || loc.getWorld() == null) return;
+
+        final int[] ticks = {0};
+        final int durationTicks = 80; // 4 seconds
+        final int period = 10;        // every 0.5 sec
+
+        Bukkit.getScheduler().runTaskTimer(plugin, task -> {
+            if (ticks[0] >= durationTicks) {
+                task.cancel();
+                return;
+            }
+            ticks[0] += period;
+
+            // spawn 1 firework each burst
+            Firework fw = loc.getWorld().spawn(loc.clone().add(0, 0.2, 0), Firework.class);
+            FireworkMeta meta = fw.getFireworkMeta();
+
+            FireworkEffect.Type type = switch (ThreadLocalRandom.current().nextInt(3)) {
+                case 0 -> FireworkEffect.Type.BALL;
+                case 1 -> FireworkEffect.Type.BALL_LARGE;
+                default -> FireworkEffect.Type.STAR;
+            };
+
+            FireworkEffect effect = FireworkEffect.builder()
+                    .with(type)
+                    .withColor(
+                            Color.RED, Color.GREEN, Color.WHITE,
+                            Color.AQUA, Color.YELLOW
+                    )
+                    .withFlicker()
+                    .withTrail()
+                    .build();
+
+            meta.clearEffects();
+            meta.addEffect(effect);
+            meta.setPower(1); // small
+            fw.setFireworkMeta(meta);
+
+            // detonate quickly so it looks like a celebration "pop" at the spot
+            Bukkit.getScheduler().runTaskLater(plugin, fw::detonate, 8L);
+
+        }, 0L, period);
+    }
+
+
     @EventHandler
     public void onZombieDeath(final EntityDeathEvent event) {
-        final Player player = event.getEntity().getKiller();
-        if (event.getEntity().getCustomName() != null && event.getEntity().getCustomName().contains("King") && event.getEntity() instanceof Zombie) {
-            event.getDrops().clear();
-            plugin.getArenaManager().getArenaByPlayer(player.getUniqueId()).endGame();
+        if (!(event.getEntity() instanceof Zombie zombie)) return;
+        if (zombie.getCustomName() == null || !zombie.getCustomName().contains("King")) return;
+
+        event.getDrops().clear();
+
+        playDragonLikeDeath(zombie.getLocation());
+
+        Player killer = zombie.getKiller();
+        if (killer != null) {
+            plugin.getArenaManager().getArenaByPlayer(killer.getUniqueId()).endGame();
         }
+
     }
+
+    private void playDragonLikeDeath(Location center) {
+        World w = center.getWorld();
+        if (w == null) return;
+
+        w.spawnParticle(Particle.EXPLOSION_HUGE, center, 1, 0, 0, 0, 0);
+        w.spawnParticle(Particle.DRAGON_BREATH, center.clone().add(0, 0.5, 0),
+                80, 1.0, 0.6, 1.0, 0.02);
+
+        final Location base = center.clone().add(0, 0.2, 0);
+
+        final int[] taskId = new int[1];
+        final int[] ticks = new int[]{0};
+        final double[] angle = new double[]{0.0};
+
+        taskId[0] = Bukkit.getScheduler().runTaskTimer(plugin, () -> {
+            ticks[0] += 2;
+            angle[0] += 0.35;
+
+            if (ticks[0] >= 80) {
+                w.spawnParticle(Particle.EXPLOSION_LARGE, base, 2, 0.2, 0.2, 0.2, 0.0);
+                w.spawnParticle(Particle.END_ROD, base.clone().add(0, 1.5, 0),
+                        18, 0.6, 0.6, 0.6, 0.02);
+                Bukkit.getScheduler().cancelTask(taskId[0]);
+                return;
+            }
+
+            double y = ticks[0] * 0.03;
+            double r = 1.1 - (ticks[0] * 0.01);
+            Location p = base.clone().add(Math.cos(angle[0]) * r, y, Math.sin(angle[0]) * r);
+
+            w.spawnParticle(Particle.PORTAL, p, 10, 0.05, 0.05, 0.05, 0.35);
+            w.spawnParticle(Particle.END_ROD, p, 2, 0.02, 0.02, 0.02, 0.0);
+
+            if (ticks[0] % 10 == 0) {
+                for (int i = 0; i < 10; i++) {
+                    Location streak = base.clone().add(
+                            ThreadLocalRandom.current().nextDouble(-2.0, 2.0),
+                            ThreadLocalRandom.current().nextDouble(0.5, 3.0),
+                            ThreadLocalRandom.current().nextDouble(-2.0, 2.0)
+                    );
+                    w.spawnParticle(Particle.END_ROD, streak, 3, 0.02, 0.02, 0.02, 0.0);
+                }
+            }
+        }, 0L, 2L).getTaskId();
+    }
+
+
+
+
 
 }
